+++ protocols/relay/src/lib.rs
@@ -45,7 +45,10 @@ pub use protocol::{HOP_PROTOCOL_NAME, STOP_PROTOCOL_NAME};
 /// Types related to the relay protocol inbound.
 pub mod inbound {
     pub mod hop {
-        pub use crate::protocol::inbound_hop::FatalUpgradeError;
+        #[deprecated(note = "Renamed to `Error`.")]
+        pub type FatalUpgradeError = Error;
+
+        pub use crate::protocol::inbound_hop::Error;
     }
 }
 
@@ -55,7 +58,7 @@ pub mod outbound {
         pub use crate::protocol::outbound_hop::{ConnectError, ProtocolViolation, ReserveError};
     }
     pub mod stop {
-        pub use crate::protocol::outbound_stop::FatalUpgradeError;
+        pub use crate::protocol::outbound_stop::{Error, ProtocolViolation};
     }
 }