+++ protocols/relay/CHANGELOG.md
@@ -1,5 +1,12 @@
 ## 0.17.0 - unreleased
 
+- Don't close connections on protocol failures within the relay-server.
+  To achieve this, error handling was restructured:
+  - `libp2p::relay::outbound::stop::FatalUpgradeError` has been removed.
+  - `libp2p::relay::outbound::stop::{Error, ProtocolViolation}` have been introduced.
+  - Several variants of `libp2p::relay::Event` have been deprecated.
+
+  See [PR 4718](https://github.com/libp2p/rust-libp2p/pull/4718).
 - Fix a rare race condition when making a reservation on a relay that could lead to a failed reservation.
   See [PR 4747](https://github.com/libp2p/rust-lib2pp/pulls/4747).
 - Propagate errors of relay client to the listener / dialer.